backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
httpx==0.28.1
gpxpy==1.6.2
pydantic-settings==2.7.0
email-validator==2.2.0
pytest==8.3.4
bcrypt==4.0.1
